+++ TransactionService/ViewModels/TransactionViewModel.cs
@@ -8,7 +8,7 @@ namespace TransactionService.ViewModels
         public Guid TransactionId { get; set; }
         public decimal Value { get; set; }
         public string? Description { get; set; }
-        public int PaymentMethodCode { get; set; }
+        public byte PaymentMethodCode { get; set; }
         public ClientViewModel Client { get; set; }
         public CreditCardViewModel CreditCard { get; set; }       
     }
